requirements.txt
pytest
httpx
websocket-client
kafka-python

locust

prometheus-api-client

fastapi
uvicorn